# default.openapi .json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Default module",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/categories": {
      "get": {
        "summary": "get-all",
        "deprecated": false,
        "description": "",
        "tags": [
          "categories"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "success": "Fetch Successfully",
                  "data": [
                    {
                      "category_id": 1,
                      "category_name": "Soy Milk",
                      "created_at": "2026-03-20T10:00:00.000Z"
                    },
                    {
                      "category_id": 2,
                      "category_name": "Coffee",
                      "created_at": "2026-03-20T10:00:00.000Z"
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "limit": 10,
                    "total": 2,
                    "total_pages": 1
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      }
    },
    "/categories/{id}": {
      "get": {
        "summary": "get-one",
        "deprecated": false,
        "description": "",
        "tags": [
          "categories"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "category_id": {
                      "type": "integer"
                    },
                    "category_name": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "category_id",
                    "category_name",
                    "created_at"
                  ]
                },
                "example": {
                  "category_id": 1,
                  "category_name": "Drinks",
                  "created_at": "2026-03-07T21:00:00Z"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/404 Not Found",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      }
    },
...
    "/menus/{id}": {
      "get": {
        "summary": "get-one",
        "deprecated": false,
        "description": "",
        "tags": [
          "menus"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "menu_id": 1,
                  "category_id": 1,
                  "menu_name": "soy_milk_light",
                  "description": "no sugar",
                  "price": 10,
                  "status": "ACTIVE",
                  "create_at": "2026-03-07T10:00:00Z",
                  "update_at": "2026-03-07T10:00:00Z"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/404 Not Found",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      },
      "put": {
        "summary": "update",
        "deprecated": false,
        "description": "",
        "tags": [
          "menus"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": false,
            "example": "application/json",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "category_id": 1,
                "menu_name": "soy_milk_light",
                "description": "less sugar",
                "price": 11,
                "status": "ACTIVE"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "message": "update successfully",
                  "data": {
                    "menu_id": 1,
                    "category_id": 1,
                    "menu_name": "soy_milk_light",
                    "description": "less sugar",
                    "price": 11,
                    "status": "ACTIVE",
                    "create_at": "2026-03-07T10:00:00Z",
                    "update_at": "2026-03-07T12:00:00Z"
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/404 Not Found",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      }
    },
    "/menus": {
      "get": {
        "summary": "get-all",
        "deprecated": false,
        "description": "",
        "tags": [
          "menus"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "example": "1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "example": "5",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "data": [
                    {
                      "menu_id": 1,
                      "category_id": 1,
                      "menu_name": "soy_milk_light",
                      "description": "no sugar",
                      "price": 10,
                      "status": "ACTIVE",
                      "create_at": "2026-03-07T10:00:00Z",
                      "update_at": "2026-03-07T10:00:00Z"
                    },
                    {
                      "menu_id": 2,
                      "category_id": 1,
                      "menu_name": "soy_milk_sweet",
                      "description": "normal sugar",
                      "price": 12,
                      "status": "ACTIVE",
                      "create_at": "2026-03-07T10:05:00Z",
                      "update_at": "2026-03-07T10:05:00Z"
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "limit": 10,
                    "total": 2,
                    "total_pages": 1
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      },
      "post": {
        "summary": "create",
        "deprecated": false,
        "description": "",
        "tags": [
          "menus"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": false,
            "example": " application/json",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "category_id": 1,
                "menu_name": "soy_milk_black_sesame",
                "description": "black sesame flavor",
                "price": 15,
                "status": "ACTIVE"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "message": "create successfully",
                  "data": {
                    "menu_id": 3,
                    "category_id": 1,
                    "menu_name": "soy_milk_black_sesame",
                    "description": "black sesame flavor",
                    "price": 15,
                    "status": "ACTIVE",
                    "create_at": "2026-03-07T11:00:00Z",
                    "update_at": "2026-03-07T11:00:00Z"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}